Refactor functionality by geometry
plugin.py was getting rather unwieldy, and there were area, cube, and position functionality spread all over, so herding them into their own folders.
